+++ render/software/src/utilities.rs
@@ -4,11 +4,11 @@ use gameplay::Angle;
 use glam::Vec2;
 
 fn player_dist_to_screen(screen_width: f32) -> f32 {
-    screen_width as f32 / 2.0 / (45.0f32.to_radians()).tan()
+    screen_width / 2.0 / (45.0f32.to_radians()).tan()
 }
 
 pub fn screen_to_x_view(x: f32, screen_width: f32) -> f32 {
-    ((screen_width as f32 / 2.0 - x) / player_dist_to_screen(screen_width)).atan()
+    ((screen_width / 2.0 - x) / player_dist_to_screen(screen_width)).atan()
 }
 
 // #[allow(clippy::excessive_precision)]
